# src/main/java/io/swagger/model/Occupation.java
package io.swagger.model;

import java.util.Objects;

import com.fasterxml.jackson.annotation.JsonProperty;

import io.swagger.annotations.ApiModelProperty;

/**
 * Occupation
 */
@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2017-04-05T13:50:22.207Z")

public class Occupation {
	@JsonProperty("name")
	private String name = null;

	@JsonProperty("type")
	private String type = null;

	@JsonProperty("scale")
	private String scale = null;

	@JsonProperty("locate")
	private String locate = null;

	public Occupation name(String name) {
		this.name = name;
		return this;
	}

	/**
	 * Get name
	 * 
	 * @return name
	 **/
	@ApiModelProperty(value = "")
	public String getName() {
		return name;
	}

	public void setName(String name) {
		this.name = name;
	}

	public Occupation type(String type) {
		this.type = type;
		return this;
	}

	/**
	 * Get type
	 * 
	 * @return type
	 **/
	@ApiModelProperty(value = "")
	public String getType() {
		return type;
	}

	public void setType(String type) {
		this.type = type;
	}

	public Occupation scale(String scale) {
		this.scale = scale;
		return this;
	}

	/**
	 * Get scale
	 * 
	 * @return scale
	 **/
	@ApiModelProperty(value = "")
	public String getScale() {
		return scale;
	}

	public void setScale(String scale) {
		this.scale = scale;
	}

	public Occupation locate(String locate) {
		this.locate = locate;
		return this;
	}

	/**
	 * Get locate
	 * 
	 * @return locate
	 **/
	@ApiModelProperty(value = "")
	public String getLocate() {
		return locate;
	}

	public void setLocate(String locate) {
		this.locate = locate;
	}

	@Override
	public boolean equals(java.lang.Object o) {
		if (this == o) {
			return true;
		}
		if (o == null || getClass() != o.getClass()) {
			return false;
		}
		Occupation occupation = (Occupation) o;
		return Objects.equals(this.name, occupation.name) && Objects.equals(this.type, occupation.type)
				&& Objects.equals(this.scale, occupation.scale) && Objects.equals(this.locate, occupation.locate);
	}

	@Override
	public int hashCode() {
		return Objects.hash(name, type, scale, locate);
	}

	@Override
	public String toString() {
		StringBuilder sb = new StringBuilder();
		sb.append("class Occupation {\n");

		sb.append("    name: ").append(toIndentedString(name)).append("\n");
		sb.append("    type: ").append(toIndentedString(type)).append("\n");
		sb.append("    scale: ").append(toIndentedString(scale)).append("\n");
		sb.append("    locate: ").append(toIndentedString(locate)).append("\n");
		sb.append("}");
		return sb.toString();
	}

	/**
	 * Convert the given object to string with each line indented by 4 spaces
	 * (except the first line).
	 */
	private String toIndentedString(java.lang.Object o) {
		if (o == null) {
			return "null";
		}
		return o.toString().replace("\n", "\n    ");
	}
}